requirements.txt
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.32.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
